+++ apps/E2E/src/MenuButtonV1/consts.ts
@@ -1,0 +1,10 @@
+export const HOMEPAGE_MENUBUTTONV1_BUTTON = 'Homepage_MenuButtonV1_Button';
+export const MENUBUTTONV1_TESTPAGE = 'MenuButtonV1_TestPage';
+
+/* E2E Testing MenuButtonV1 1 */
+export const MENUBUTTONV1_TEST_COMPONENT = 'MenuButtonV1_Test_Component'; // A component on each specific test page
+export const MENUBUTTONV1_ACCESSIBILITY_LABEL = 'E2E testing MenuButtonV1 accessibility label';
+
+/* E2E Testing MenuButtonV1 2 */
+export const MENUBUTTONV1_NO_A11Y_LABEL_COMPONENT = 'MenuButtonV1_No_A11y_label_Component';
+export const MENUBUTTONV1_TEST_COMPONENT_LABEL = 'Test MenuButtonV1 2 - No Accessibility Label'; // A component on each specific test page