+++ apps/E2E/src/TabsV1/pages/TabsV1PageObject.ts
@@ -1,0 +1,39 @@
+import { TABSITEMV1_TEST_COMPONENT, TABSV1_TESTPAGE, TABSV1_TEST_COMPONENT, HOMEPAGE_TABSV1_BUTTON } from '../consts';
+import { BasePage, By } from '../../common/BasePage';
+
+class TabsV1PageObject extends BasePage {
+  /******************************************************************/
+  /**************** UI Element Interaction Methods ******************/
+  /******************************************************************/
+  async getTabItemAccesibilityRole(): Promise<string> {
+    return await (await this._tabItem).getAttribute('ControlType');
+  }
+
+  /*****************************************/
+  /**************** Getters ****************/
+  /*****************************************/
+  get _testPage() {
+    return By(TABSV1_TESTPAGE);
+  }
+
+  get _pageName() {
+    return TABSV1_TESTPAGE;
+  }
+
+  get _primaryComponent() {
+    return By(TABSV1_TEST_COMPONENT);
+  }
+
+  get _pageButton() {
+    return By(HOMEPAGE_TABSV1_BUTTON);
+  }
+
+  /***********/
+  /* TabItem *
+  /***********/
+  get _tabItem() {
+    return By(TABSITEMV1_TEST_COMPONENT);
+  }
+}
+
+export default new TabsV1PageObject();