+++ apps/fluent-tester/src/TestComponents/Button/deprecated/E2EButtonTest.tsx
@@ -10,7 +10,8 @@ import {
   BUTTON_NO_A11Y_LABEL_COMPONENT_DEPRECATED,
   BUTTON_ACCESSIBILITY_LABEL_DEPRECATED,
   BUTTON_TEST_COMPONENT_LABEL_DEPRECATED,
-} from '../../../../../E2E/src/Button/consts';
+} from '../../../../../E2E/src/ButtonLegacy/consts';
+import { testProps } from '../../Common/TestProps';
 
 export const E2EButtonTest_deprecated: React.FunctionComponent = () => {
   const [buttonPressed, setButtonPressed] = React.useState(false);
@@ -24,12 +25,25 @@ export const E2EButtonTest_deprecated: React.FunctionComponent = () => {
       <Stack style={stackStyle}>
         <Button
           content="This is a button for E2E testing"
-          testID={BUTTON_TEST_COMPONENT_DEPRECATED}
           onClick={onClick}
           accessibilityLabel={BUTTON_ACCESSIBILITY_LABEL_DEPRECATED}
+          /* For Android E2E testing purposes, testProps must be passed in after accessibilityLabel. */
+          {...testProps(BUTTON_TEST_COMPONENT_DEPRECATED)}
         />
-        <Button content={BUTTON_TEST_COMPONENT_LABEL_DEPRECATED} testID={BUTTON_NO_A11Y_LABEL_COMPONENT_DEPRECATED} onClick={onClick} />
-        {buttonPressed ? <Text testID={BUTTON_ON_PRESS_DEPRECATED}>Button Pressed</Text> : null}
+        <Button
+          content={BUTTON_TEST_COMPONENT_LABEL_DEPRECATED}
+          onClick={onClick}
+          /* For Android E2E testing purposes, testProps must be passed in after accessibilityLabel. */
+          {...testProps(BUTTON_NO_A11Y_LABEL_COMPONENT_DEPRECATED)}
+        />
+        {buttonPressed ? (
+          <Text
+            /* For Android E2E testing purposes, testProps must be passed in after accessibilityLabel. */
+            {...testProps(BUTTON_ON_PRESS_DEPRECATED)}
+          >
+            Button Pressed
+          </Text>
+        ) : null}
       </Stack>
     </View>
   );