+++ apps/fluent-tester/src/TestComponents/CheckboxV1/E2ECheckboxV1Test.tsx
@@ -1,0 +1,53 @@
+import { Checkbox } from '@fluentui-react-native/experimental-checkbox';
+import * as React from 'react';
+import { View } from 'react-native';
+import { stackStyle } from '../Common/styles';
+import {
+  CHECKBOXV1_TEST_COMPONENT,
+  CHECKBOXV1_ACCESSIBILITY_LABEL,
+  CHECKBOXV1_NO_A11Y_LABEL_COMPONENT,
+  CHECKBOXV1_TEST_COMPONENT_LABEL,
+  CHECKBOXV1_ON_PRESS,
+} from '../../../../E2E/src/CheckboxV1/consts';
+import { Stack } from '@fluentui-react-native/stack';
+import { Text } from '@fluentui/react-native';
+import { testProps } from '../Common/TestProps';
+
+export const E2ECheckboxV1Test: React.FunctionComponent = () => {
+  const [checkboxPressed, setCheckboxPressed] = React.useState(false);
+
+  const onClick = React.useCallback(
+    (_e, checked) => {
+      setCheckboxPressed(checked);
+    },
+    [setCheckboxPressed],
+  );
+
+  return (
+    <View>
+      <Stack style={stackStyle}>
+        <Checkbox
+          accessibilityLabel={CHECKBOXV1_ACCESSIBILITY_LABEL}
+          disabled={false}
+          label="Testing accessibilityLabel"
+          onChange={onClick}
+          /* For Android E2E testing purposes, testProps must be passed in after accessibilityLabel. */
+          {...testProps(CHECKBOXV1_TEST_COMPONENT)}
+        />
+        <Checkbox
+          label={CHECKBOXV1_TEST_COMPONENT_LABEL}
+          /* For Android E2E testing purposes, testProps must be passed in after accessibilityLabel. */
+          {...testProps(CHECKBOXV1_NO_A11Y_LABEL_COMPONENT)}
+        />
+        {checkboxPressed ? (
+          <Text
+            /* For Android E2E testing purposes, testProps must be passed in after accessibilityLabel. */
+            {...testProps(CHECKBOXV1_ON_PRESS)}
+          >
+            Checkbox Selected
+          </Text>
+        ) : null}
+      </Stack>
+    </View>
+  );
+};