+++ apps/fluent-tester/src/TestComponents/LinkV1/InlineLinksTest.tsx
@@ -1,0 +1,36 @@
+import * as React from 'react';
+import { Alert } from 'react-native';
+import { LinkV1 as Link, TextV1 as Text } from '@fluentui/react-native';
+import { Stack } from '@fluentui-react-native/stack';
+import { stackStyle } from '../Common/styles';
+
+export const InlineLinks: React.FunctionComponent = () => {
+  const doPress = React.useCallback(() => Alert.alert('Alert.', 'You have been alerted.'), []);
+  const doAllyTap = React.useCallback(() => Alert.alert('Alert.', 'You have invoked onAllyTap.'), []);
+
+  return (
+    <Stack style={stackStyle}>
+      <Text>
+        Click{' '}
+        <Link inline onPress={doPress} onAccessibilityTap={doAllyTap}>
+          this link
+        </Link>{' '}
+        to alert me.
+      </Text>
+      <Text>
+        This{' '}
+        <Link inline onPress={doPress} disabled focusable>
+          link
+        </Link>{' '}
+        is disabled but focusable.
+      </Text>
+      <Text>
+        Follow this{' '}
+        <Link inline url="https://www.bing.com/">
+          link
+        </Link>{' '}
+        to navigate.
+      </Text>
+    </Stack>
+  );
+};