+++ apps/fluent-tester/src/TestComponents/TabsV1/index.ts
@@ -1,0 +1,1 @@
+export * from './TabsV1Test';