+++ packages/experimental/RadioGroup/CHANGELOG.md
@@ -1,8 +1,24 @@
 # Change Log - @fluentui-react-native/experimental-radio-group
 
-This log was last generated on Fri, 30 Dec 2022 16:09:27 GMT and should not be manually modified.
+This log was last generated on Tue, 03 Jan 2023 23:38:40 GMT and should not be manually modified.
 
 <!-- Start content -->
+
+## 0.9.1
+
+Tue, 03 Jan 2023 23:38:40 GMT
+
+### Patches
+
+- Bump @fluentui-react-native/text to v0.19.7
+
+## 0.9.0
+
+Tue, 03 Jan 2023 17:37:38 GMT
+
+### Minor changes
+
+- adding additional snapshot tests (gulnazsayed@microsoft.com)
 
 ## 0.8.13